# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Automation Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AccountStatus = (typeof AccountStatus)[keyof typeof AccountStatus];

export const AccountStatus = {
  active: "active",
  inactive: "inactive",
  error: "error",
} as const;

export interface Account {
  id: number;
  name: string;
  email: string;
  password: string;
  twofaSecret?: string;
  status: AccountStatus;
  createdAt: string;
}

export type CreateAccountBodyStatus =
  (typeof CreateAccountBodyStatus)[keyof typeof CreateAccountBodyStatus];

export const CreateAccountBodyStatus = {
  active: "active",
  inactive: "inactive",
  error: "error",
} as const;

export interface CreateAccountBody {
  name: string;
  email: string;
  password: string;
  twofaSecret?: string;
  status?: CreateAccountBodyStatus;
}

export type BotStatus = (typeof BotStatus)[keyof typeof BotStatus];

export const BotStatus = {
  active: "active",
  inactive: "inactive",
  error: "error",
} as const;

export interface Bot {
  id: number;
  accountId?: number;
  name: string;
  token: string;
  prefix: string;
  status: BotStatus;
  createdAt: string;
}

export type CreateBotBodyStatus =
  (typeof CreateBotBodyStatus)[keyof typeof CreateBotBodyStatus];

export const CreateBotBodyStatus = {
  active: "active",
  inactive: "inactive",
  error: "error",
} as const;

export interface CreateBotBody {
  accountId?: number;
  name: string;
  token: string;
  prefix: string;
  status?: CreateBotBodyStatus;
}

export type SessionStatus = (typeof SessionStatus)[keyof typeof SessionStatus];

export const SessionStatus = {
  running: "running",
  waiting: "waiting",
  stopped: "stopped",
  completed: "completed",
  error: "error",
} as const;

export interface Session {
  id: number;
  status: SessionStatus;
  total: number;
  progress: number;
  waitSeconds: number;
  logs?: string;
  createdAt: string;
  updatedAt: string;
}

export interface SessionOrNull {
  session?: Session;
  exists: boolean;
}

export interface CreateSessionBody {
  total: number;
  waitSeconds: number;
}

export type UpdateSessionBodyStatus =
  (typeof UpdateSessionBodyStatus)[keyof typeof UpdateSessionBodyStatus];

export const UpdateSessionBodyStatus = {
  running: "running",
  waiting: "waiting",
  stopped: "stopped",
  completed: "completed",
  error: "error",
} as const;

export interface UpdateSessionBody {
  status?: UpdateSessionBodyStatus;
  progress?: number;
  logs?: string;
}

export interface AutomationRules {
  createTeams: boolean;
  createBots: boolean;
  linkBots: boolean;
  quantity: number;
  botPrefix: string;
  waitMinutes: number;
}

export interface AiChatBody {
  message: string;
  context?: string;
}

export interface AiChatResponse {
  reply: string;
  suggestions?: string[];
}

export interface Stats {
  totalAccounts: number;
  activeAccounts: number;
  totalBots: number;
  activeBots: number;
  totalSessions: number;
  lastSessionStatus?: string;
}

export interface DeleteResult {
  success: boolean;
}