# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Automation Studio API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all saved accounts
 */
export const ListAccountsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  password: zod.string(),
  twofaSecret: zod.string().optional(),
  status: zod.enum(["active", "inactive", "error"]),
  createdAt: zod.string(),
});
export const ListAccountsResponse = zod.array(ListAccountsResponseItem);

/**
 * @summary Add a new Discord account
 */
export const CreateAccountBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  password: zod.string(),
  twofaSecret: zod.string().optional(),
  status: zod.enum(["active", "inactive", "error"]).optional(),
});

/**
 * @summary Get a single account
 */
export const GetAccountParams = zod.object({
  id: zod.coerce.number(),
});

export const GetAccountResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  password: zod.string(),
  twofaSecret: zod.string().optional(),
  status: zod.enum(["active", "inactive", "error"]),
  createdAt: zod.string(),
});

/**
 * @summary Update an account
 */
export const UpdateAccountParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateAccountBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  password: zod.string(),
  twofaSecret: zod.string().optional(),
  status: zod.enum(["active", "inactive", "error"]).optional(),
});

export const UpdateAccountResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  password: zod.string(),
  twofaSecret: zod.string().optional(),
  status: zod.enum(["active", "inactive", "error"]),
  createdAt: zod.string(),
});

/**
 * @summary Delete an account
 */
export const DeleteAccountParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteAccountResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary List all bots
 */
export const ListBotsResponseItem = zod.object({
  id: zod.number(),
  accountId: zod.number().optional(),
  name: zod.string(),
  token: zod.string(),
  prefix: zod.string(),
  status: zod.enum(["active", "inactive", "error"]),
  createdAt: zod.string(),
});
export const ListBotsResponse = zod.array(ListBotsResponseItem);

/**
 * @summary Create a new bot record
 */
export const CreateBotBody = zod.object({
  accountId: zod.number().optional(),
  name: zod.string(),
  token: zod.string(),
  prefix: zod.string(),
  status: zod.enum(["active", "inactive", "error"]).optional(),
});

/**
 * @summary Get a single bot
 */
export const GetBotParams = zod.object({
  id: zod.coerce.number(),
});

export const GetBotResponse = zod.object({
  id: zod.number(),
  accountId: zod.number().optional(),
  name: zod.string(),
  token: zod.string(),
  prefix: zod.string(),
  status: zod.enum(["active", "inactive", "error"]),
  createdAt: zod.string(),
});

/**
 * @summary Update a bot (e.g. reset token, rename)
 */
export const UpdateBotParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateBotBody = zod.object({
  accountId: zod.number().optional(),
  name: zod.string(),
  token: zod.string(),
  prefix: zod.string(),
  status: zod.enum(["active", "inactive", "error"]).optional(),
});

export const UpdateBotResponse = zod.object({
  id: zod.number(),
  accountId: zod.number().optional(),
  name: zod.string(),
  token: zod.string(),
  prefix: zod.string(),
  status: zod.enum(["active", "inactive", "error"]),
  createdAt: zod.string(),
});

/**
 * @summary Delete a bot
 */
export const DeleteBotParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteBotResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary List all automation sessions
 */
export const ListSessionsResponseItem = zod.object({
  id: zod.number(),
  status: zod.enum(["running", "waiting", "stopped", "completed", "error"]),
  total: zod.number(),
  progress: zod.number(),
  waitSeconds: zod.number(),
  logs: zod.string().optional(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListSessionsResponse = zod.array(ListSessionsResponseItem);

/**
 * @summary Start a new automation session
 */
export const CreateSessionBody = zod.object({
  total: zod.number(),
  waitSeconds: zod.number(),
});

/**
 * @summary Get the currently active session (if any)
 */
export const GetActiveSessionResponse = zod.object({
  session: zod
    .object({
      id: zod.number(),
      status: zod.enum(["running", "waiting", "stopped", "completed", "error"]),
      total: zod.number(),
      progress: zod.number(),
      waitSeconds: zod.number(),
      logs: zod.string().optional(),
      createdAt: zod.string(),
      updatedAt: zod.string(),
    })
    .optional(),
  exists: zod.boolean(),
});

/**
 * @summary Get a single session
 */
export const GetSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetSessionResponse = zod.object({
  id: zod.number(),
  status: zod.enum(["running", "waiting", "stopped", "completed", "error"]),
  total: zod.number(),
  progress: zod.number(),
  waitSeconds: zod.number(),
  logs: zod.string().optional(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update session (progress, status, logs)
 */
export const UpdateSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateSessionBody = zod.object({
  status: zod
    .enum(["running", "waiting", "stopped", "completed", "error"])
    .optional(),
  progress: zod.number().optional(),
  logs: zod.string().optional(),
});

export const UpdateSessionResponse = zod.object({
  id: zod.number(),
  status: zod.enum(["running", "waiting", "stopped", "completed", "error"]),
  total: zod.number(),
  progress: zod.number(),
  waitSeconds: zod.number(),
  logs: zod.string().optional(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Stop an active session
 */
export const StopSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const StopSessionResponse = zod.object({
  id: zod.number(),
  status: zod.enum(["running", "waiting", "stopped", "completed", "error"]),
  total: zod.number(),
  progress: zod.number(),
  waitSeconds: zod.number(),
  logs: zod.string().optional(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Get the current automation rules
 */
export const GetRulesResponse = zod.object({
  createTeams: zod.boolean(),
  createBots: zod.boolean(),
  linkBots: zod.boolean(),
  quantity: zod.number(),
  botPrefix: zod.string(),
  waitMinutes: zod.number(),
});

/**
 * @summary Update automation rules
 */
export const UpdateRulesBody = zod.object({
  createTeams: zod.boolean(),
  createBots: zod.boolean(),
  linkBots: zod.boolean(),
  quantity: zod.number(),
  botPrefix: zod.string(),
  waitMinutes: zod.number(),
});

export const UpdateRulesResponse = zod.object({
  createTeams: zod.boolean(),
  createBots: zod.boolean(),
  linkBots: zod.boolean(),
  quantity: zod.number(),
  botPrefix: zod.string(),
  waitMinutes: zod.number(),
});

/**
 * @summary Send a message to the AI assistant for automation guidance
 */
export const AiChatBody = zod.object({
  message: zod.string(),
  context: zod.string().optional(),
});

export const AiChatResponse = zod.object({
  reply: zod.string(),
  suggestions: zod.array(zod.string()).optional(),
});

/**
 * @summary Get dashboard stats (totals, active bots, session counts)
 */
export const GetStatsResponse = zod.object({
  totalAccounts: zod.number(),
  activeAccounts: zod.number(),
  totalBots: zod.number(),
  activeBots: zod.number(),
  totalSessions: zod.number(),
  lastSessionStatus: zod.string().optional(),
});